refactor: export testing-utils from integrations utils
