+++ test/integration/bin/multi-path/index.test.ts
@@ -1,6 +1,5 @@
 import { readFile } from 'fs/promises'
-import { createIntegrationTest } from '../../utils'
-import { existsFile } from '../../../testing-utils'
+import { createIntegrationTest, existsFile } from '../../utils'
 
 describe('integration', () => {
   test(`bin/multi-path`, async () => {
