+++ test/integration/entry-index-index/index.test.ts
@@ -1,5 +1,4 @@
-import { createIntegrationTest } from '../utils'
-import { assertFilesContent } from '../../testing-utils'
+import { createIntegrationTest, assertFilesContent } from '../utils'
 
 describe('integration', () => {
   test(`entry-index-index`, async () => {
